refactor : BonusNumber 추가 / 테스트코드 추가 필요
