+++ src/main/java/lotto/domain/BonusNumber.java
@@ -1,0 +1,12 @@
+package lotto.domain;
+
+public class BonusNumber extends LottoNumber {
+
+    public BonusNumber(int value) {
+        super(value);
+    }
+
+    public boolean isSameNumber(final LottoNumber number) {
+        return this.getValue() == number.getValue();
+    }
+}
