+++ src/main/java/lotto/domain/result/WinningTicket.java
@@ -1,19 +1,20 @@
 package lotto.domain.result;
 
+import lotto.domain.BonusNumber;
 import lotto.domain.LottoNumber;
 import lotto.domain.ticket.LottoTicket;
 
 public class WinningTicket {
     private final LottoTicket lottoTicket;
-    private final LottoNumber bonusNumber;
+    private final BonusNumber bonusNumber;
 
-    public WinningTicket(LottoTicket lottoTicket, LottoNumber bonusNumber) {
+    public WinningTicket(LottoTicket lottoTicket, BonusNumber bonusNumber) {
         validate(lottoTicket, bonusNumber);
         this.lottoTicket = lottoTicket;
         this.bonusNumber = bonusNumber;
     }
 
-    private void validate(LottoTicket lottoTicket, LottoNumber bonusNumber) {
+    private void validate(LottoTicket lottoTicket, BonusNumber bonusNumber) {
         if (lottoTicket.contains(bonusNumber)) {
             throw new IllegalArgumentException("보너스번호와 당첨번호는 중복될 수 없습니다");
         }
@@ -24,6 +25,6 @@ public class WinningTicket {
     }
 
     public boolean matchBonus(LottoNumber lottoNumber) {
-        return bonusNumber.equals(lottoNumber);
+        return bonusNumber.isSameNumber(lottoNumber);
     }
 }
